requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pandas
PyPDF2
requests
